requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
ffmpeg-python==0.2.0
Pillow==10.1.0
loguru==0.7.2
aiohttp==3.9.1 